es6 module loader integration
